Adds sessions last ts
