+++ backend/app/api/platform/endpoints/projects.py
@@ -46,15 +46,12 @@ from app.services.mongo.projects import (
     delete_project_from_id,
     delete_project_related_resources,
     email_project_tasks,
-    get_all_sessions,
     get_project_by_id,
     update_project,
 )
-from app.services.mongo.search import (
-    search_sessions_in_project,
-    search_tasks_in_project,
-)
+
 from app.services.mongo.tasks import get_all_tasks
+from app.services.mongo.sessions import get_all_sessions
 from app.services.mongo.users import fetch_users_metadata
 from app.services.slack import slack_notification
 from app.services.universal_loader.universal_loader import universal_loader
@@ -183,31 +180,6 @@ async def get_events(
     propelauth.require_org_member(user, project.org_id)
     events = await get_all_events(project_id=project_id, limit=limit)
     return Events(events=events)
-
-
-@router.post(
-    "/projects/{project_id}/search/tasks",
-    response_model=SearchResponse,
-    description="Perform a semantic search in the project's sessions",
-)
-async def post_search_tasks(
-    project_id: str,
-    search_query: SearchQuery,
-    user: User = Depends(propelauth.require_user),
-):
-    """
-    Get the resulting session_ids of a semantic search in the project's sessions.
-    The search is based on embedding similarity of the text conversation to the query.
-    """
-
-    project = await get_project_by_id(project_id)
-    propelauth.require_org_member(user, project.org_id)
-    # Perform the semantic search
-    relevant_tasks = await search_tasks_in_project(
-        project_id=project_id,
-        search_query=search_query.query,
-    )
-    return SearchResponse(task_ids=[task.id for task in relevant_tasks])
 
 
 @router.get(
