+++ backend/app/api/v2/endpoints/projects.py
@@ -12,10 +12,8 @@ from app.api.v2.models import (
     Tasks,
 )
 from app.security import authenticate_org_key, verify_propelauth_org_owns_project_id
-from app.services.mongo.projects import (
-    backcompute_recipes,
-    get_all_sessions,
-)
+from app.services.mongo.projects import backcompute_recipes
+from app.services.mongo.sessions import get_all_sessions
 from app.services.mongo.tasks import (
     fetch_flattened_tasks,
     get_all_tasks,
